+++ wrapper/include/pyfmu/utils.hpp
